+++ app/grandchallenge/teams/models.py
@@ -4,7 +4,7 @@ from django.core.exceptions import ValidationError, PermissionDenied
 from django.db import models
 
 from grandchallenge.challenges.models import Challenge
-from grandchallenge.core.urlresolvers import reverse
+from grandchallenge.subdomains.urls import reverse
 
 
 class Team(models.Model):
