+++ app/tests/datasets_tests/test_views.py
@@ -1,7 +1,7 @@
 import pytest
 from django.core.exceptions import ValidationError
 
-from grandchallenge.core.urlresolvers import reverse
+from grandchallenge.subdomains.urls import reverse
 from grandchallenge.datasets.models import ImageSet, AnnotationSet
 from tests.cases_tests.test_background_tasks import (
     create_raw_upload_image_session
