+++ app/tests/participants_tests/test_permissions.py
@@ -1,6 +1,6 @@
 import pytest
 
-from grandchallenge.core.urlresolvers import reverse
+from grandchallenge.subdomains.urls import reverse
 from tests.factories import RegistrationRequestFactory
 from tests.utils import (
     validate_admin_only_view,
